fix warning on 32 bit compilation

+++ src/string.cc
@@ -113,7 +113,7 @@ String to_string(int val)
 String to_string(size_t val)
 {
     char buf[16];
-    sprintf(buf, "%lu", val);
+    sprintf(buf, "%lu", (unsigned long)val);
     return buf;
 }
